Engine core and IBM Watson extraction
