+++ README.md
@@ -40,7 +40,7 @@ The following pieces of middleware are currently available;
 
 | Name | Description | NPM |
 | ---- | ----------- | ------- |
-| [@botbuildercommunity/text-analytics-middleware](libraries/botbuilder-text-analytics-middleware/README.md) | Use Cognitive Services Text Analytics API for sentiment analysis, language detection, key phrases and entity extraction. | [![NPM Version](https://img.shields.io/npm/v/@botbuildercommunity/text-analytics-middleware.svg)](https://www.npmjs.com/package/@botbuildercommunity/text-analytics-middleware) |
+| [@botbuildercommunity/middleware-text-analytics](libraries/botbuilder-middleware-text-analytics/README.md) | Use Cognitive Services Text Analytics API for sentiment analysis, language detection, key phrases and entity extraction. | [![NPM Version](https://img.shields.io/npm/v/@botbuildercommunity/middleware-text-analytics.svg)](https://www.npmjs.com/package/@botbuildercommunity/middleware-text-analytics) |
 | [@botbuildercommunity/spell-check-middleware](libraries/botbuilder-spell-check-middleware/README.md) | Use Cognitive Services Spell Check API to detect misspellings and correct these. | [![NPM Version](https://img.shields.io/npm/v/@botbuildercommunity/spell-check-middleware.svg)](https://www.npmjs.com/package/@botbuildercommunity/spell-check-middleware) |
 | [@botbuildercommunity/text-recognizer-middleware](libraries/botbuilder-text-recognizer-middleware/README.md) | Use the [Microsoft Recognizers Text](https://github.com/microsoft/Recognizers-Text) Suite for recognizing certain text sequences. | [![NPM Version](https://img.shields.io/npm/v/@botbuildercommunity/text-recognizer-middleware.svg)](https://www.npmjs.com/package/@botbuildercommunity/text-recognizer-middleware) |
