+++ libraries/botbuilder-middleware-engine-core/src/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @module botbuildercommunity/middleware-engine-core
+ */
+
+export * from "./engine";
