+++ libraries/botbuilder-middleware-text-analytics/.npmignore
@@ -1,0 +1,1 @@
+node_modules/
