+++ libraries/botbuilder-middleware-text-analytics/README.md
@@ -4,23 +4,23 @@ The Text Analytics Middleware offers Bot Framework middleware components for the
 
 ## Installing
 
-    npm install @botbuildercommunity/text-analytics-middleware --save
+    npm install @botbuildercommunity/middlware-text-analytics --save
 
 ## Usage
 
 All middleware is created and used in the same way. For example, for sentiment analysis, import the `SentimentAnalysis` class from the package, and add it to your bot adapter:
 
 ```typescript
-import { SentimentAnalysis } from '@botbuildercommunity/text-analytics-middleware';
+import { SentimentAnalysis } from '@botbuildercommunity/middleware-text-analytics';
 
-adapter.use(new SentimentAnalysis(YOUR_TEXT_ANALYTICS_KEY, TEXT_ANALYTICS_API_ENDPOINT, IEngineOptions));
+adapter.use(new SentimentAnalysis(YOUR_TEXT_ANALYTICS_KEY, TEXT_ANALYTICS_API_ENDPOINT, SERVICE_CLIENT_OPTIONS));
 ```
 
-The `IEngineOptions` parameter is optional. See the next section for details.
+The `CLIENT_OPTIONS` parameter is optional.
 
 > Note that the TEXT_ANALYTICS_API_ENDPOINT will be the Cognitive Services endpoint root. For example: https://eastus.api.cognitive.microsoft.com
 
-When used, the `turnState` on the `TurnContext` will have a property named `sentimentScore` between 0 and 1. A full example can be seen in the [`bot.js`](test/bot.js) bot test file.
+When used, the `turnState` on the `TurnContext` will have a property named `sentimentScore` between 0 and 1. A full example can be seen in the [`app-cs.js`](example/app-cs.js) bot test file.
 
 Supported middleware classes include:
 
@@ -37,7 +37,3 @@ In each case, properties are added to the `turnState` of the `TurnContext`. You 
 * `context.turnState.get('language')` //This is a string for `LanguageDetection`
 * `context.turnState.get('keyPhrases')` //This is an array of strings for `KeyPhrases`
 * `context.turnState.get('textEntities')` //This is an array of `EntityRecord` types` for `EntityExtraction`
-
-## Engine and Engine Options
-
-This package supports two engines: Azure's Cognitive Services Text Analytics API
