+++ libraries/botbuilder-middleware-text-analytics/example/botbuilder-sentiment-middleware-testbot.bot
