+++ libraries/botbuilder-middleware-text-analytics/example/package.json
