+++ libraries/botbuilder-middleware-text-analytics/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@botbuildercommunity/text-analytics-middleware",
-  "version": "1.0.0",
+  "name": "@botbuildercommunity/middleware-text-analytics",
+  "version": "0.9.0",
   "description": "Bot Framework middleware component for the Cognitive Services Text Analytics API, including sentiment analysis, language detection, entity extraction, and keyword recognition.",
   "main": "./lib/index.js",
   "typings": "./lib/index.d.ts",
@@ -34,7 +34,6 @@
   },
   "homepage": "https://github.com/BotBuilderCommunity/botbuilder-community-js#readme",
   "dependencies": {
-    "watson-developer-cloud": "^4.0.1",
     "azure-cognitiveservices-textanalytics": "^2.1.0",
     "botbuilder": "^4.5.3",
     "ms-rest-azure": "^2.6.0"
