+++ libraries/botbuilder-middleware-text-analytics/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @module botbuildercommunity/middleware-text-analytics
+ */
+
+export * from "./engine";
+export * from "./sentiment";
+export * from "./languageDetection";
+export * from "./keyPhrases";
+export * from "./entity";
