+++ libraries/botbuilder-middleware-text-analytics/src/sentiment.ts
@@ -1,0 +1,35 @@
+import { Middleware, TurnContext, ActivityTypes } from "botbuilder";
+import { Engine } from "../../botbuilder-middleware-engine-core/lib/engine";
+import { CognitiveServiceEngine } from "./engine";
+
+/**
+ * @module botbuildercommunity/middleware-text-analytics
+ */
+
+export class SentimentAnalysis implements Middleware {
+    public engine: Engine;
+    constructor(public serviceKey: string, public endpoint: string, public options?: any) {
+        this.engine = new CognitiveServiceEngine(serviceKey, endpoint, options);
+    }
+    public async onTurn(context: TurnContext, next: () => Promise<void>) {
+        if(context.activity.type === ActivityTypes.Message) {
+            const input = {
+                documents: [
+                    {
+                        "id": "1"
+                        , "text": context.activity.text
+                    }
+                ]
+            };
+            try {
+                const result = await this.engine.sentiment(input);
+                const s = result.documents[0].score;
+                context.turnState.set("sentimentScore", s);
+            }
+            catch(e) {
+                throw new Error(`Failed to process sentiment on ${context.activity.text}. Error: ${e}`);
+            }
+        }
+        await next();
+    }
+}
