+++ libraries/botbuilder-middleware-watson-nlu/example/app-watson.js
