+++ libraries/botbuilder-middleware-watson-nlu/example/botbuilder-emotion-middleware-testbot.bot
@@ -1,0 +1,15 @@
+{
+    "name": "botbuilder-emotion-middleware-testbot",
+    "description": "",
+    "secretKey": "",
+    "services": [
+        {
+            "appId": "",
+            "id": "23838dc0-c7cc-11e8-8398-fb2384adc22d",
+            "type": "endpoint",
+            "appPassword": "",
+            "endpoint": "http://localhost:3978/api/messages",
+            "name": "http://localhost:3978/api/messages"
+        }
+    ]
+}
