+++ libraries/botbuilder-middleware-watson-nlu/src/keyPhrases.ts
@@ -1,16 +1,16 @@
 import { Middleware, TurnContext, ActivityTypes } from "botbuilder";
-import { CognitiveServicesCredentials } from "ms-rest-azure";
-import { TextAnalyticsClient } from "azure-cognitiveservices-textanalytics";
-import { Engine } from "./engine";
+import { Engine } from "../../botbuilder-middleware-engine-core/lib/engine";
+import { WatsonEngine } from "./engine";
 
 /**
- * @module botbuildercommunity/text-analytics
+ * @module botbuildercommunity/middleware-watson-nlu
  */
 
-export class LanguageDetection implements Middleware {
+export class KeyPhrases implements Middleware {
     public engine: Engine;
+
     constructor(public serviceKey: string, public endpoint: string, public options?: any) {
-        this.engine = Engine.getEngine(serviceKey, endpoint, options);
+        this.engine = new WatsonEngine(serviceKey, endpoint, options);
     }
     public async onTurn(context: TurnContext, next: () => Promise<void>) {
         if(context.activity.type === ActivityTypes.Message) {
@@ -23,12 +23,12 @@ export class LanguageDetection implements Middleware {
                 ]
             };
             try {
-                const result = await this.engine.detectLanguage(input);
-                const l = result.documents[0].detectedLanguages;
-                context.turnState.set("language", l);
+                const result = await this.engine.keyPhrases(input);
+                const k = result.documents[0].keyPhrases;
+                context.turnState.set("keyPhrases", k);
             }
             catch(e) {
-                throw new Error(`Failed to process language on ${context.activity.text}. Error: ${e}`);
+                throw new Error(`Failed to process key phrases on ${context.activity.text}. Error: ${e}`);
             }
         }
         await next();
