+++ libraries/botbuilder-middleware-watson-nlu/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "target": "ES2015",
+    "module": "commonjs",
+    "declaration": true,
+    "sourceMap": true,
+    "outDir": "./lib",
+    "rootDir": "./src",
+    "types" : ["node"]
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
